Move Rate class into dedicated file
