+++ src/client/components/robot/model.ts
@@ -1,6 +1,6 @@
-import { throttle } from 'throttle-debounce'
 import { observable } from 'mobx'
-import { runInAction } from 'mobx'
+
+import { Rate } from './rate'
 
 type RobotModelOpts = {
   id: string
@@ -36,41 +36,4 @@ export class RobotModel {
   static of(opts: RobotModelOpts) {
     return new RobotModel(opts)
   }
-}
-
-type RateOpts = {
-  smoothing: number
-  unitTime: number
-}
-
-class Rate {
-  @observable rate = 0
-  private currentRate: number = 0
-  private smoothing: number
-  private unitTime: number
-  private lastUpdateTime?: number
-
-  constructor({ smoothing, unitTime }: RateOpts) {
-    this.smoothing = smoothing
-    this.unitTime = unitTime
-  }
-
-  update(newCount: number) {
-    if (this.lastUpdateTime) {
-      this.currentRate =
-        this.currentRate * this.smoothing +
-        newCount / (Date.now() - this.lastUpdateTime) *
-        (1.0 - this.smoothing)
-
-      this.commit(this.currentRate)
-    }
-
-    this.lastUpdateTime = Date.now()
-  }
-
-  private commit = throttle(1000, (rate: number) => {
-    runInAction(() => {
-      this.rate = rate * this.unitTime
-    })
-  })
 }
